Initial Commit
So, to start, you'll want to uploat the file 'lcdTest.ino' to the arduino, and then run the fasterAPI.py script! if you want to change the coin that is being monitored, just open up the python script, then change the name on line 51!
The circuit diagram is in 'image.png' for y'all. i suggest usign a breadboard for the connections otherwise it can get a bit confusing to fix if a wire comes out!